Fixes for the comments
